requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
pandas==2.0.3
SQLAlchemy==2.0.15
cx_Oracle==8.3.0
impyla==0.18.0
python-dotenv==1.0.0